Render recipient form with context
