+++ src/app/(with-sidebar)/recipients/components/send/send-to-dialog.stories.tsx
@@ -4,6 +4,7 @@ import { fn } from '@storybook/test';
 
 import { mockRecipients } from '@/api/recipients/mock';
 import { Button } from '@/components/ui/button';
+import { sleep } from '@/lib/utils';
 
 import { SendToDialog } from './send-to-dialog';
 
@@ -14,19 +15,10 @@ const meta = {
     layout: 'centered',
   },
   args: {
-    onSend: fn(),
+    onSend: fn(async () => {
+      await sleep(3000);
+    }),
     onAdd: fn(),
-  },
-} satisfies Meta<typeof SendToDialog>;
-
-export default meta;
-
-type Story = StoryObj<typeof SendToDialog>;
-
-export const Default: Story = {
-  args: {
-    maxAmount: 1000,
-    recipients: mockRecipients,
     open: true,
   },
   render: (args) => {
@@ -40,5 +32,23 @@ export const Default: Story = {
         <SendToDialog {...args} open={open} setOpen={setOpen} />
       </>
     );
+  },
+} satisfies Meta<typeof SendToDialog>;
+
+export default meta;
+
+type Story = StoryObj<typeof SendToDialog>;
+
+export const Default: Story = {
+  args: {
+    maxAmount: 1000,
+    recipients: mockRecipients,
+  },
+};
+
+export const WithRecipient: Story = {
+  args: {
+    ...Default.args,
+    selectedRecipientId: mockRecipients[0].id,
   },
 };
